Crate problem096

Problem 96 (Su Doku)

Solution detail

Ignoring the boilerplate of reading in the puzzles from file, the meat of this problem is in quickly finding the solution to a given Sudoku puzzle.

Depth-first search is a prime candidate for this - given a puzzle to solve, fill in the empty cells in order, branching on all digits that are allowed to go into that cell, and backtracking when an impossibility is encountered.

After solving all of the puzzles, adding up to get the solution is a trivial matter.

Structs

SudokuTree
The state of the current puzzle during a depth-first solve

SudokuTreeStep
A description of a step that can be taken in the search tree.

Functions

answer
Solve the problem, returning the answer as a String

main

read_grid

solve
Solve all of the given Sudoku puzzles and sum the three-digit numbers found in the top-left of each solved puzzle.

solve_sudoku
Find the solution to a single Sudoku puzzle

Type Definitions

Coords

Grid